requirements.txt
asn1crypto==0.22.0
azure-common==1.1.8
azure-nspkg==2.0.0
azure-storage==0.35.1
certifi==2017.7.27.1
cffi==1.10.0
chardet==3.0.4
cryptography==2.0.3
idna==2.5
oauthlib==2.0.2
pycparser==2.18
pydocumentdb==2.2.0
python-dateutil==2.6.1
requests==2.18.3
requests-oauthlib==0.8.0
six==1.10.0
tweepy==3.5.0
urllib3==1.22